requirements.txt
hydra-core==1.3.2
omegaconf==2.3.0
openai==2.14.0
python-dotenv==1.2.1
PyYAML==6.0.3
